♻️ refactor: consolidate all headers under includes/
Move bvh_internal.h and bvhd_internal.h back to includes/spatial/.
All headers now live exclusively under includes/ regardless of
public/internal scope, keeping a single authoritative location.
