+++ includes/spatial/bvh_internal.h
