+++ includes/spatial/bvhd_internal.h
